fix(dts-generator): add @types discovery to generate check-compile path
TS6 no longer auto-includes @types packages when the compiler is invoked
programmatically. The previous fix (631f1de) only addressed runCheck.ts
but the generation path in generate.ts has the same problem — causing
"Cannot find namespace 'THREE'" errors for sap.ui.vk in the legacy-free
build.

Extracted the @types discovery logic into a shared utility
(utils/discover-types.ts) and applied it to both generate.ts and
runCheck.ts. Also aligned generate.ts module settings to ESNext +
Bundler to prevent bare-specifier rejection of ambient declare module
declarations.
