+++ packages/dts-generator/src/generate-from-paths.ts
@@ -56,15 +56,9 @@ export interface GenerateFromPathsConfig {
   dependenciesDTSPathForCheck: string;
 
   /**
-   * Comma-separated list of package names of the libraries on which the currently to-be-built types depends.
-   *
-   * This is meant for entire npm packages developed separately (often by others), not sibling libraries built in the same batch.
-   * E.g. when a custom UI5 control library is built by an application team, then it usually depends on the OpenUI5 types because
-   * those define the base classes like Control.
-   * Setting this has the effect that for the TS compilation check, the `types` field of the package.json file will be set to the
-   * respective package names and any other type packages are no longer considered.
-   *
-   * Only needed for the check.
+   * @deprecated Since 4.0.3. Declared @types/* dependencies are now auto-discovered
+   * from the nearest package.json. This option is still accepted for backward
+   * compatibility and merges additively with the discovered types.
    */
   dependenciesTypePackagesForCheck?: string;
